Add dashboard variables to single-instance built-in dashboards
Adds a dashboard variable to single-instance built-in dashboards so that
an instance is always selected and the dashboard will render something
meaningful. This will allow us to make links to built-in dashboards on
the dashboards page point to a dashboard page rather than the catalog,
which would be consistent with the rest of the dashboard links on that
page.

This commit only adds a dashboard variable to "single-instance"
dashboards, or dashboards whose `sf_discoverySelectors` contained only
clauses that began with `"_exists_"`. The dimension of the variable was
chosen as the value of that `"_exists_"` selector if there was only one
selector or the value which seemed most likely to represent the relevant
instance if there were multiple selectors.

The code to add to each serialized dashboard was determined by exporting
the same dashboard with and without dashboard variables and diff-ing the
resulting JSON files:

    <     "variables" : [ {
    <       "dimension" : "host",
    <       "alias" : "host",
    <       "value" : "i-c3dae4e8",
    <       "globalScope" : true,
    <       "required" : true,
    <       "description" : "",
    <       "restricted" : false,
    <       "preferredSuggestions" : [ ]
    <     } ]
    ---
    >     "variables" : [ ]